+++ src/collective/wfeffectiverange/vocabulary.py
@@ -13,79 +13,49 @@ _pmf = MessageFactory('plone')
 
 
 @implementer(IContextSourceBinder)
-class TransitionsSource(object):
+class BaseTransitionsSource(object):
 
-    def __init__(self, fieldname, transition=None, portal_type=None):
-        self.fieldname = fieldname
-        self.transition = transition
-        self.portal_type = portal_type
+    FIELD_NAME = ''
 
-    def __call__(self, context):
-        wftool = api.portal.get_tool('portal_workflow')
+    def __init__(self):
+        self.transition = None
+        self.portal_type = None
+
+    def _init_call(self, context):
         from .behaviors import WFEffectiveRange
         if isinstance(context, WFEffectiveRange):
             context = context.context
         url = context.REQUEST.getURL()
         addform = '++add++' in url
         addtranslationform = '++addtranslation++' in url
-
+        self.add = addform or addtranslationform
         if self.portal_type is None:
             if addform:
                 # todo: could be done in one step with re.match.
                 # strip the /@@validate_field!
-                url = re.sub('\/@{2}.*', '', url)
+                self.url = re.sub('\/@{2}.*', '', url)
                 # get portal_type from addform
                 self.portal_type = re.split('.*\+{2}add\+{2}', url)[1]
             elif addtranslationform:
-                url = re.sub('\/@{2}.*', '', url)
+                self.url = re.sub('\/@{2}.*', '', url)
                 # get portal_type from addform
                 self.portal_type = re.split(
                     '.*\+{2}addtranslation\+{2}', url
                 )[1]
             else:
                 self.portal_type = context.portal_type
+        wftool = api.portal.get_tool('portal_workflow')
+        self.workflows = wftool.getWorkflowsFor(self.portal_type)
+        self.submitted = context.REQUEST.form.get(self.FIELD_NAME, None)
+        return context
 
-        wfs = wftool.getWorkflowsFor(self.portal_type)
-        if len(wfs) == 0:
-            return SimpleVocabulary([])
-        elif len(wfs) > 1:
-            raise ValueError(
-                _(u'Multiple workflows are not supported.')
-            )
-        wf = wfs[0]
-
-        # if no effective_transition is set get all possible transitions
-        # for the expires_transition
-        # if an effective_transition is set, only get the allowed
-        # transitions for that
-        try:
-            cet = context.effective_transition
-        except AttributeError:
-            cet = None
-        if self.transition is None \
-                and not addform \
-                and not addtranslationform \
-                and cet is not None \
-                and self.fieldname == 'expires_transition':
-            self.transition = context.effective_transition
-
-        if self.transition and self.transition != '--NOVALUE--':
-            state = wf.transitions[self.transition].new_state_id
-
-            # get current state for portal_type
-            # If it is given as a string it returns the default state.
-            # see PLIP 217 Workflow by adaptation
-        elif addform or addtranslationform:
-            state = api.content.get_state(self.portal_type)
-        else:
-            state = api.content.get_state(context)
-
-        # get possible transitions for this state
-        transitions = wf.states[state].transitions
+    def _vocab_transitions(self, state):
+        """get possible transitions for this state
+        """
+        transitions = self.workflow.states[state].transitions
         terms = []
-
         for transition in transitions:
-            trans_id = wf.transitions[transition].id
+            trans_id = self.workflow.transitions[transition].id
             terms.append(
                 SimpleVocabulary.createTerm(
                     trans_id,
@@ -93,5 +63,68 @@ class TransitionsSource(object):
                     _pmf(trans_id)
                 )
             )
-
         return SimpleVocabulary(terms)
+
+
+class EffectiveTransitionSource(BaseTransitionsSource):
+
+    def __call__(self, context):
+        context = self._init_call(context)
+        if len(self.workflows) == 0:
+            return SimpleVocabulary([])
+        elif len(self.workflows) > 1:
+            raise ValueError(
+                _(u'Multiple workflows are not supported.')
+            )
+        self.workflow = self.workflows[0]
+        if self.add:
+            # get current state for portal_type
+            # If it is given as a string it returns the default state.
+            # see PLIP 217 Workflow by adaptation
+            state = api.content.get_state(self.portal_type)
+        else:
+            state = api.content.get_state(context)
+        return self._vocab_transitions(state)
+
+
+class ExpiresTransitionSource(BaseTransitionsSource):
+
+    FIELD_NAME = 'form.widgets.IWFEffectiveRange.effective_transition'
+
+    def __init__(self, transition=None, portal_type=None):
+        self.transition = transition
+        self.portal_type = portal_type
+
+    def __call__(self, context):
+        context = self._init_call(context)
+        if len(self.workflows) == 0:
+            return SimpleVocabulary([])
+        elif len(self.workflows) > 1:
+            raise ValueError(
+                _(u'Multiple workflows are not supported.')
+            )
+        self.workflow = self.workflows[0]
+
+        if self.transition == u'--NOVALUE--':
+            self.transition = None
+
+        if self.transition is None:
+            if self.submitted:
+                self.transition = self.submitted[0]
+            elif not self.add:
+                self.transition = getattr(
+                    context,
+                    'effective_transition',
+                    None
+                )
+
+        if self.transition:
+            state = self.workflow.transitions[self.transition].new_state_id
+        elif self.add:
+            # get current state for portal_type
+            # If it is given as a string it returns the default state.
+            # see PLIP 217 Workflow by adaptation
+            state = api.content.get_state(self.portal_type)
+        else:
+            state = api.content.get_state(context)
+        return self._vocab_transitions(state)